+++ packages/table/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @grid-table/view
 
+## 0.6.49
+
+### Patch Changes
+
+- feat:添加cell 延迟渲染
+- Updated dependencies
+  - @grid-table/core@0.6.49
+  - @grid-table/basic@0.6.49
+
+## 0.6.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/core@0.6.48
+  - @grid-table/basic@0.6.48
+
+## 0.6.47
+
+### Patch Changes
+
+- fixed:调整滚动方案
+- Updated dependencies
+  - @grid-table/core@0.6.47
+  - @grid-table/basic@0.6.47
+
 ## 0.6.46
 
 ### Patch Changes